added events for mouse click on selection fields in the UI, as well as events for mouse click on game field segments. NOTE: The second one should be updated when the Tower classes are created.
Furthermore, refactored the Geometry.Point structure as now the x and y of it are double rather than integers. Fixed the corresponding constructors for all classes dependent on this class.

+++ TowerDefenseGame/TowerDefenseGame/Models/Enemies/BasicEnemy.cs
@@ -8,7 +8,7 @@ namespace TowerDefenseGame.Models.Enemies
         private const int LifePoints = 50;
         private const int Speed = 1;
 
-        public BasicEnemy(int x, int y) : base(x, y, 50, 50, BasicEnemy.LifePoints, BasicEnemy.Speed, Brushes.Black)
+        public BasicEnemy(double x, double y) : base(x, y, 50, 50, BasicEnemy.LifePoints, BasicEnemy.Speed, Brushes.Black)
         {
 
         }
